requirements.txt
wheel
djangorestframework
django-filter
drf-yasg
djangorestframework-simplejwt==4.4.0
python-decouple
Pillow
psycopg2-binary
python-slugify
celery
redis
django-cors-headers
Django==3.2.5